llm/requirements.txt
kafka-python==2.1.3
pandas
psycopg2-binary==2.9.9